+++ src/identifiers/tests.rs
@@ -4,10 +4,10 @@ pub(crate) fn read_samples_from_txt(filename: &str) -> Vec<String> {
     let txt = format!("{}/testdata/{}", env!("CARGO_MANIFEST_DIR"), filename);
     read_to_string(txt)
         .unwrap()
-        .split("\n")
+        .split('\n')
         .flat_map(|line| {
             let line = line.trim().to_string();
-            if line.starts_with("#") || line.is_empty() {
+            if line.starts_with('#') || line.is_empty() {
                 None
             } else {
                 Some(line)
@@ -18,7 +18,7 @@ pub(crate) fn read_samples_from_txt(filename: &str) -> Vec<String> {
 
 pub(crate) fn apply_to_samples_from_txt<F>(filename: &str, f: F)
 where
-    F: Fn(&str) -> (),
+    F: Fn(&str),
 {
     for sample in read_samples_from_txt(filename) {
         // dbg!(&sample);