+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+* Improve date parsing, switch to new chrono `NaiveDate::from_ymd_opt` and `NaiveTime::from_hms_opt` APIs.
+
 ## [0.1.0] - 2022-11-28
 
 Initial release